requirements.txt
Django==1.11
djangorestframework==3.7.7
psycopg2==2.6.2
pytz==2018.3
